breaking: Moves getTransactionsStatus one level up from wtIndex to libsIndex
