+++ test/utils/scripts/localtestnet.sh
@@ -24,7 +24,6 @@ testrpc_running() {
 }
 
 start_testrpc() {
-  # We define 10 accounts with balance lots of ether, needed for high-value tests.
   local accounts=(
     --account="0xe8280389ca1303a2712a874707fdd5d8ae0437fab9918f845d26fd9919af5a92,10000000000000000000000000000000000000000000000000000000000000000000000000000000"
     --account="0xed095a912033d26dc444d2675b33414f0561af170d58c33f394db8812c87a764,10000000000000000000000000000000000000000000000000000000000000000000000000000000"
