+++ test/wt-libs/data-model/web3-json/hotel.spec.js
@@ -30,7 +30,7 @@ describe('WTLibs.data-model.web3-json.hotel', () => {
 
     it('should throw when we want hotel from a bad address', async () => {
       try {
-        const hotelProvider = await HotelDataProvider.createInstance(dataModel.commonWeb3Utils, dataModel.commonWeb3Contracts, await indexDataProvider.__getDeployedIndex(), '0x96eA4BbF71FEa3c9411C1Cefc555E9d7189695fA');
+        const hotelProvider = await HotelDataProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex(), '0x96eA4BbF71FEa3c9411C1Cefc555E9d7189695fA');
         await hotelProvider.__getContractInstance();
         throw new Error('should not have been called');
       } catch (e) {
@@ -40,7 +40,7 @@ describe('WTLibs.data-model.web3-json.hotel', () => {
 
     it('should throw when we want hotel without an address', async () => {
       try {
-        const hotelProvider = await HotelDataProvider.createInstance(dataModel.commonWeb3Utils, dataModel.commonWeb3Contracts, await indexDataProvider.__getDeployedIndex());
+        const hotelProvider = await HotelDataProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex());
         await hotelProvider.__getContractInstance();
         throw new Error('should not have been called');
       } catch (e) {
@@ -50,7 +50,7 @@ describe('WTLibs.data-model.web3-json.hotel', () => {
 
     it('should throw if we try to get data from network in a hotel without address', async () => {
       try {
-        const hotelProvider = await HotelDataProvider.createInstance(dataModel.commonWeb3Utils, dataModel.commonWeb3Contracts, await indexDataProvider.__getDeployedIndex());
+        const hotelProvider = await HotelDataProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex());
         await hotelProvider.ethBackedDataset.__syncRemoteData();
         throw new Error('should not have been called');
       } catch (e) {
@@ -60,7 +60,7 @@ describe('WTLibs.data-model.web3-json.hotel', () => {
 
     it('should cache contract instances', async () => {
       assert.equal(getHotelContractSpy.callCount, 0);
-      const hotelProvider = await HotelDataProvider.createInstance(dataModel.commonWeb3Utils, dataModel.commonWeb3Contracts, await indexDataProvider.__getDeployedIndex(), '0x4a763f50dfe5cf4468b4171539e021a26fcee0cc');
+      const hotelProvider = await HotelDataProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex(), '0x4a763f50dfe5cf4468b4171539e021a26fcee0cc');
       assert.equal(getHotelContractSpy.callCount, 1);
       await hotelProvider.__getContractInstance();
       assert.equal(getHotelContractSpy.callCount, 1);
@@ -78,7 +78,7 @@ describe('WTLibs.data-model.web3-json.hotel', () => {
         manager: '0xd39ca7d186a37bb6bf48ae8abfeb4c687dc8f906',
       });
 
-      const hotelProvider = await HotelDataProvider.createInstance(dataModel.commonWeb3Utils, dataModel.commonWeb3Contracts, await indexDataProvider.__getDeployedIndex(), result.address);
+      const hotelProvider = await HotelDataProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex(), result.address);
       const newName = 'Random changed name';
       hotelProvider.name = newName;
       await hotelProvider.updateOnNetwork({
@@ -86,7 +86,7 @@ describe('WTLibs.data-model.web3-json.hotel', () => {
         to: indexDataProvider.address,
       });
       assert.equal(await hotelProvider.name, newName);
-      let freshHotelProvider = await HotelDataProvider.createInstance(dataModel.commonWeb3Utils, dataModel.commonWeb3Contracts, await indexDataProvider.__getDeployedIndex(), result.address);
+      let freshHotelProvider = await HotelDataProvider.createInstance(dataModel.web3Utils, dataModel.web3Contracts, await indexDataProvider.__getDeployedIndex(), result.address);
       assert.equal(await hotelProvider.name, await freshHotelProvider.name);
       // And remove the hotel to keep the data consistent
       await indexDataProvider.removeHotel(freshHotelProvider);
